requirements/test.txt
pytest
pytest-xdist[psutil]
python-dotenv
